thesis/requirements.txt
baukit @ git+https://github.com/davidbau/baukit@20a3c05d505bbf598b336db16e2d1ecee4ae7466
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
diffusers==0.12.1
filelock==3.9.0
fonttools==4.38.0
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
tokenizers==0.13.2
torch
torchaudio
torchvision
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.4.0
urllib3==1.26.14
zipp==3.13.0
scipy
accelerate